minor files restructuring
